backend/ingest/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==42.0.8
Django==5.2
django-filter==24.3
djangorestframework==3.16.0
drf-yasg==1.21.10
Faker==37.1.0
idna==3.10
inflection==0.5.1
Markdown==3.8
mysql-connector==2.2.9
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
